requirements.txt
python-telegram-bot==20.7
tvdatafeed==2.1.0
yfinance==0.2.33
pandas==2.1.4
pandas-ta==0.3.14b0
mplfinance==0.12.10b0
APScheduler==3.10.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
matplotlib==3.8.2
Pillow==10.1.0
